requirements.txt
Flask==3.0.2
google-generativeai>=0.3.0
Pillow>=10.0.0
python-dotenv>=1.0.0